Fix standalone toolchain to pass all tests with libstdc++
The main idea was to copy armeabi-v7a libs to the place where armeabi
libs were placed. GCC to not need these copies but clangs fails to link
even when it links armeabi-v7a programs.

Also clang was not able to find correct bits/c++config.h which is placed
into abi dependent subdirectory for GCC.